# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP Basic Authentication. Click \"Authorize\" button and enter username/password configured in the server.",
        "title": "MID Bootstrap Server API",
        "contact": {
            "name": "MAG Support",
            "url": "https://github.com/your-org/mid-bootstrap-server"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.4.3"
    },
    "basePath": "/",
    "paths": {
...
        "/api/approve": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Approves a pending bootstrap request and generates a token for the agent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Approve a bootstrap request",
                "parameters": [
                    {
                        "description": "Approval details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ApprovalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approval confirmation with token info",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request or not pending",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Request not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Token generation failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/audit": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns audit log entries with optional filtering by hostname or event type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get audit log",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of entries to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by hostname",
                        "name": "hostname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by event type (approval, denial, token_delivered, etc.)",
                        "name": "event_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of audit entries",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.AuditEntry"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/deny": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Denies a pending bootstrap request with a reason",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Deny a bootstrap request",
                "parameters": [
                    {
                        "description": "Denial details with reason",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.DenialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Denial confirmation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Request not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/generate-token": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Generates a bootstrap token that operators can use with the agent's local /bootstrap endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Generate bootstrap token for manual onboarding",
                "parameters": [
                    {
                        "description": "Token generation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ApprovalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token details including the actual token value",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Request not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Token generation failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/manual-bootstrap": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Generates a bootstrap token for a machine without requiring a bootstrap request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Generate manual bootstrap token",
                "parameters": [
                    {
                        "description": "Manual bootstrap request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.ManualBootstrapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated bootstrap token",
                        "schema": {
                            "$ref": "#/definitions/types.BootstrapToken"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to generate token",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/requests": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns bootstrap requests, optionally filtered by status (supports pagination)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List bootstrap requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status (pending, approved, denied)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit for pagination (default 50, max 500, 0 for all)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of bootstrap requests",
                        "schema": {
                            "$ref": "#/definitions/store.PaginatedRequests"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/stats": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns statistics about bootstrap requests (total, pending, approved, denied)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get dashboard statistics",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics",
                        "schema": {
                            "$ref": "#/definitions/types.DashboardStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/system-stats": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns statistics about registered agents (total, active, CA ready)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Get system statistics",
                "responses": {
                    "200": {
                        "description": "System statistics",
                        "schema": {
                            "$ref": "#/definitions/types.SystemStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/systems": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns registered agents with their status information (supports pagination)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "List registered agents",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit for pagination (default 50, max 500, 0 for all)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of registered agents",
                        "schema": {
                            "$ref": "#/definitions/store.PaginatedRegistrations"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/vault-status": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns the health status of the connected Vault server",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get Vault health status",
                "responses": {
                    "200": {
                        "description": "Vault health status",
                        "schema": {
                            "$ref": "#/definitions/vault.VaultHealthStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "server.ManualBootstrapRequest": {
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "types.ApprovalRequest": {
            "type": "object",
            "properties": {
                "approved_by": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "types.DenialRequest": {
            "type": "object",
            "properties": {
                "denied_by": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}